+++ pre-install.sh
@@ -2,15 +2,68 @@
 
 set -euo pipefail
 
-echo -e "\e[34mInstalling Python 2 and 3 Setting the default Python to version 3.\e[0m"
-sudo apt-get update
-sudo apt-get install -y python2.7 python3.8
-sudo update-alternatives --install /usr/bin/python python /usr/bin/python2.7 1
-sudo update-alternatives --install /usr/bin/python python /usr/bin/python3.8 2
-sudo update-alternatives  --set python /usr/bin/python3.8
+UPDATE=false
+SHOW_HELP=false
+VERBOSE=false
+while [[ $# -gt 0 ]]; do
+  key="$1"
+  case $key in
+    --update|-u)
+    UPDATE=true
+    shift
+    ;;
+    --help|-h)
+    SHOW_HELP=true
+    break
+    ;;
+    --verbose)
+    VERBOSE=true
+    shift
+    ;;
+    *)
+    shift
+    ;;
+  esac
+done
 
-echo -e "\e[34mUpgrading all packages.\e[0m"
-sudo apt-get upgrade -y
+if $SHOW_HELP; then
+  cat <<EOF
+Packages pre installer.
 
-echo -e "\e[34mSetting default time zone to São Paulo.\e[0m"
-sudo ln -fs /usr/share/zoneinfo/America/Sao_Paulo /etc/localtime
+Usage:
+  `readlink -f $0` [flags]
+
+Flags:
+  -u, --update             Will download and install/reinstall even if the tools are already installed
+      --verbose            Show verbose output
+  -h, --help               help
+EOF
+  exit 0
+fi
+
+if $VERBOSE; then
+  echo Running `basename "$0"` $ALL_ARGS
+  echo Update is $UPDATE
+fi
+
+if ! hash python2.7 2>/dev/null || ! hash python3.8 2>/dev/null; then
+  echo -e "\e[34mInstalling Python 2 and 3.\e[0m"
+  sudo apt-get update
+  sudo apt-get install -y python2.7 python3.8
+fi
+if ! update-alternatives --display python &>/dev/null; then
+  echo -e "\e[34mSetting the default Python to version 3.\e[0m"
+  sudo update-alternatives --install /usr/bin/python python /usr/bin/python2.7 1
+  sudo update-alternatives --install /usr/bin/python python /usr/bin/python3.8 2
+  sudo update-alternatives  --set python /usr/bin/python3.8
+fi
+
+if $UPDATE; then
+  echo -e "\e[34mUpgrading all packages.\e[0m"
+  sudo apt-get upgrade -y
+fi
+
+if ! [ -L /etc/localtime ] || [ `realpath /etc/localtime` != "/usr/share/zoneinfo/America/Sao_Paulo" ]; then
+  echo -e "\e[34mSetting default time zone to São Paulo.\e[0m"
+  sudo ln -fs /usr/share/zoneinfo/America/Sao_Paulo /etc/localtime
+fi
